backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
alembic>=1.14.0
pydantic-settings>=2.7.0
yfinance>=0.2.50
gspread>=6.1.0
apscheduler>=3.10.0
boto3>=1.35.0
google-auth>=2.0.0
PyJWT>=2.0.0
